requirements.txt
bcrypt==4.0.1
capstone==5.0.0rc2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colored-traceback==0.3.0
cryptography==39.0.0
idna==3.4
intervaltree==3.1.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
paramiko==3.0.0
plumbum==1.8.1
psutil==5.9.4
pwntools==4.9.0
pycparser==2.21
pyelftools==0.29
Pygments==2.14.0
PyNaCl==1.5.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.2
ROPGadget==7.2
rpyc==5.3.0
six==1.16.0
sortedcontainers==2.4.0
unicorn==2.0.1.post1
urllib3==1.26.14
